revert back to old hashing, variable not being found

+++ csrc/table.c
@@ -107,10 +107,10 @@ void tableAddAll(Table* from, Table* to) {
     }
 }
 ObjString* tableFindString(Table* table, const char* chars,
-                           int length, uint64_t hash) {
+                           int length, uint32_t hash) {
     if (table->count == 0) return NULL;
 
-    uint64_t index = hash & (table->capacity - 1);
+    uint32_t index = hash & (table->capacity - 1);
     for (;;) {
         Entry* entry = &table->entries[index];
         if (entry->key == NULL) {
